% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacc-package.R
\docType{package}
\name{cacc-package}
\alias{cacc}
\alias{cacc-package}
\title{cacc: cacc: Central America climate change}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains geospatial data of Central America to do climate change analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://manuelspinola.github.io/cacc/}
}

}
\keyword{internal}
