% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacc.R
\name{cacc}
\alias{cacc}
\title{Conjunctive Analysis of Case Configurations}
\usage{
cacc(data, x, y)
}
\arguments{
\item{data}{A data frame or a tibble.}

\item{x}{A vector of names of the independent variables, without quotes. Variables must be categorical, either integer, character, or factor.}

\item{y}{Name of the dependent variable, without quotes. Variable must be dichotomous with values 0 (absence) and 1 (presence).}
}
\value{
Returns a tibble with the CACC matrix.
}
\description{
Computes a Conjunctive Analysis of Case Configurations (CACC).
}
\examples{
cacc(data = test_data, x = c(iv1, iv2, iv3, iv4), y = dv1)
cacc(data = test_data, x = c(iv1_c, iv2_c, iv3_c, iv4_c), y = dv1)
}
