% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onharassment}
\alias{onharassment}
\title{Profiles of 4174 Spanish students}
\format{
A data frame with 4174 rows and 12 variables:
\describe{
\item{sex}{\code{Factor}. The students' self-reported sex.}
\item{age}{\code{Factor}. The students' self-reported age}
\item{hours}{\code{Factor}. The students' self-reported number of daily hours spent online.}
\item{snapchat}{\code{Factor}. Whether students report using the social media Snapchat on a daily basis.}
\item{instagram}{\code{Factor}. Whether students report using the social media Instagram on a daily basis.}
\item{facebook}{\code{Factor}. Whether students report using the social media Facebook on a daily basis.}
\item{twitter}{\code{Factor}. Whether students report using the social media Twitter on a daily basis.}
\item{name}{\code{Factor}. Whether students report using their real names on social media.}
\item{photos}{\code{Factor}. Whether students report uploading personal photos to social media.}
\item{privacy}{\code{Factor}. Whether students report restricting their social media access to contacts only.}
\item{rep_victim}{\code{Factor}. Whether students report repeatedly suffering online harassment.}
\item{rep_offender}{\code{Factor}. Whether students report repeatedly committing online harassment.}
}
}
\source{
Moneva, A., Miró-Llinares, F., & Hart, T. C. (2021). Hunter or Prey? Exploring the situational profiles that define repeated online harassment victims and offenders. \emph{Deviant Behavior, 42}(11), 1366-1381. https://doi.org/10.1080/01639625.2020.1746135
}
\usage{
onharassment
}
\description{
A dataset containing the demographics, online routines, and self-reported online harassment repeat victimization and offending of 4174 Spanish non-university education students.
}
\keyword{datasets}
