% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect.R
\name{plot_effect}
\alias{plot_effect}
\title{Density Plot for the Main Effect}
\usage{
plot_effect(cacc_matrix, variable, value)
}
\arguments{
\item{cacc_matrix}{A tibble. The output of the \code{cacc} function.}

\item{variable}{A single variable name contained in a \code{cacc_matrix}.}

\item{value}{A single numeric or character value the \code{variable} specified can take.}
}
\value{
Returns a ggplot object.
}
\description{
Plots an annotated boxplot and kernel density estimate to visualize the distribution of the main effect that a specific value of a variable produces on the outcome probability in a \code{cacc_matrix}.
}
\examples{
plot_effect(
  cacc_matrix = cacc(data = test_data, ivs = c(iv1, iv2, iv3, iv4), dv = dv1),
  variable = iv4,
  value = 0
)

}
