% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data to compute a Conjunctive Analysis of Case Configurations}
\format{
A data frame with 100 rows and 9 variables:
\describe{
\item{iv1, iv2, iv3, iv4}{Example integer independent variables.}
\item{iv1_c, iv2_c, iv3_c, iv4_c}{Example character or factor independent variables.}
\item{dv1}{Example integer, dichotomous, dependent variable.}
}
}
\usage{
test_data
}
\description{
A dataset containing synthetic example data to conduct a Conjunctive Analysis of Case Configurations.
}
\keyword{datasets}
