% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacc.R
\name{cacc}
\alias{cacc}
\title{Conjunctive Analysis of Case Configurations}
\usage{
cacc(data, ivs, dv)
}
\arguments{
\item{data}{A data frame or a tibble.}

\item{ivs}{A vector of names of the independent variables, without quotes. Variables must be categorical, either integer, character, or factor.}

\item{dv}{Name of the dependent variable, without quotes. Variable must be dichotomous with values 0 (absence) and 1 (presence).}
}
\value{
Returns a tibble with the CACC matrix.
}
\description{
Computes a Conjunctive Analysis of Case Configurations (CACC).
}
\examples{
cacc(data = test_data, ivs = c(iv1, iv2, iv3, iv4), dv = dv1)
cacc(data = test_data, ivs = c(iv1_c, iv2_c, iv3_c, iv4_c), dv = dv1)

# Syntax with the native R pipe
test_data |> cacc(ivs = c(iv1, iv2, iv3, iv4), dv = dv1)
}
\references{
Miethe, T. D., Hart, T. C., & Regoeczi, W. C. (2008). The Conjunctive Analysis of Case Configurations: An Exploratory Method for Discrete Multivariate Analyses of Crime Data. \emph{Journal of Quantitative Criminology, 24}, 227–241. https://doi.org/10.1007/s10940-008-9044-8
}
