% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacc.R
\name{cacc}
\alias{cacc}
\title{Conjunctive Analysis of Case Configurations}
\usage{
cacc(data, x, y)
}
\arguments{
\item{data}{Input data. A data frame or a tibble.}

\item{x}{Concatenated names of the independent variables without quotes.}

\item{y}{Name of the dependent variable without quotes.}
}
\value{
Returns a tibble with the CACC matrix.
}
\description{
Computes a Conjunctive Analysis of Case Configurations (CACC).
}
\examples{
cacc(data = test_data, x = c(iv1, iv2, iv3, iv4), y = dv1)
}
